requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes --format requirements-txt
-e .
autopep8==2.3.2
    # via encryptocli
cffi==2.0.0
    # via
    #   cryptography
    #   encryptocli
click==8.3.1
    # via typer
colorama==0.4.6
    # via
    #   click
    #   crayons
    #   encryptocli
crayons==0.4.0
    # via
    #   encryptocli
    #   stegano
cryptography==46.0.3
    # via encryptocli
inquirerpy==0.3.4
    # via encryptocli
markdown-it-py==4.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
numpy==2.2.6 ; python_full_version < '3.11'
    # via
    #   encryptocli
    #   opencv-python
numpy==2.4.1 ; python_full_version >= '3.11'
    # via
    #   encryptocli
    #   opencv-python
opencv-python==4.13.0.90
    # via
    #   encryptocli
    #   stegano
pfzy==0.3.4
    # via inquirerpy
piexif==1.1.3
    # via
    #   encryptocli
    #   stegano
pillow==11.3.0
    # via
    #   encryptocli
    #   stegano
prompt-toolkit==3.0.52
    # via inquirerpy
pycodestyle==2.14.0
    # via
    #   autopep8
    #   encryptocli
pycparser==2.23
    # via
    #   cffi
    #   encryptocli
pyfiglet==1.0.4
    # via encryptocli
pygments==2.19.2
    # via
    #   encryptocli
    #   rich
python-gnupg==0.5.3
    # via encryptocli
regex==2026.1.15
    # via encryptocli
rich==14.3.1
    # via typer
shellingham==1.5.4
    # via typer
six==1.17.0
    # via encryptocli
stegano==2.1.0
    # via encryptocli
termcolor==3.3.0
    # via encryptocli
toml==0.10.2
    # via encryptocli
tomli==2.4.0 ; python_full_version < '3.11'
    # via autopep8
typer==0.21.1
    # via encryptocli
typing-extensions==4.15.0
    # via
    #   cryptography
    #   typer
wcwidth==0.2.14
    # via
    #   encryptocli
    #   prompt-toolkit
